Adds download option for videos
This commit adds a context menu entry for downloading video streams to a
folder specified in the plug-in configuration. It was inspired by the
YouTube plug-in (http://wiki.xbmc.org/index.php?title=Add-on:YouTube),
which uses the SimpleDownloader plug-in
(http://wiki.xbmc.org/index.php?title=Add-on:Simple_Downloader_for_xbmc_plugins)
for downloading video streams.

The new dependency to the SimpleDownloader is an optional one. The code
reports an error to the user, if the dependency is not
installed/activated.

Translations for the new menu entries are given in German and English.

Changes tested with XBMC 12.2 on Windows and Linux.