Refactoring for having a single paring function
To fix some of the parsing bugs like in #2, a single parsing function
for determining the attributes of a video would be easier. Thus, this
changes move all the parsing from the "list*" methods to an
"iterateContent" (parsing list of content elements) and
"extractVideoDescription" (attributes of a single content element).